refactor(gui): split the four pages out of flow.ts, and the chart out of trends.ts
flow.ts was 3,728 lines holding four unrelated things. The Energy Overview turned out to depend on exactly
three of them — the moment picker and the two functions that phrase its query and describe its answer — so
those move to history-control.ts and the board moves to its own file. Two pages building the same control
is how they end up describing the same window differently, which is the reason it is one module rather
than a copy each.

trends.ts kept the bar chart it draws with. That is a drawing problem, not an energy one, and the rules in
it are worth finding: a day with no reading is an empty slot and never a zero-height bar, a negative value
belongs below the zero line rather than cancelling a positive one, and the hover target is the whole day
column because a stacked segment can be a pixel tall.

No behaviour change: 757 tests and all 12 GUI checks pass unchanged, including the layout and sankey checks
that render the pages that moved.

flow.ts 3728 -> 3121, trends.ts 521 -> 380, plus charts.ts, energy-board.ts and history-control.ts.
