+++ rPDU2MQTT.Web/web/build.mjs
@@ -24,6 +24,8 @@ const MODULES = [
   'realtime.ts',
   'dirty.ts',
   'energy.ts',
+  'history-control.ts',
+  'charts.ts',
   'palette.ts',
   'overrides.ts',
   'sections/paths.ts',
@@ -31,6 +33,7 @@ const MODULES = [
   'sections/control.ts',
   'sections/livedata.ts',
   'sections/flow.ts',
+  'sections/energy-board.ts',
   'sections/nodedata.ts',
   'sections/trends.ts',
   'sections/export.ts',
