+++ rPDU2MQTT.Web/web/src/history-control.ts
@@ -1,0 +1,162 @@
+// Choosing a moment to look at, shared by every view that can show one.
+//
+// Its own module because two pages build it — the Sankey and the Energy board — and a control with two
+// implementations is a control that answers differently depending on which page you are standing on. What
+// it produces is the query (`at`, `span`) and the sentence describing what came back, so those two things
+// are decided once as well.
+import { el, btn } from '../helpers.js';
+import { state } from '../state.js';
+
+/// The `at`/`span` part of a flow query, and the sentence that says what came back.
+///
+/// Both pages ask the same question and must describe the answer the same way — a diagram or a board that
+/// looks live while showing last Tuesday is the worst thing either page can do.
+export function historyQuery(hist: { at: () => string, span: () => number }): string {
+  const at = hist.at();
+  if (!at) return '';
+  const span = hist.span();
+  return '&at=' + encodeURIComponent(at) + (span > 1 ? '&span=' + span : '');
+}
+
+export function historyNote(body: any): string {
+  if (!body || !body.historical) return '';
+  const when = new Date(body.at).toLocaleString();
+  const days = Number(body.spanDays) || 1;
+  const what = days > 1 ? `${days} days to ${new Date(body.at).toLocaleDateString()}` : when;
+  // A window with days missing from it is not that window. Name the nodes rather than quietly showing a
+  // short total as a whole one.
+  const short = (body.incomplete || []) as { node: string, days: number }[];
+  const gap = short.length
+    ? ` · incomplete: ${short.slice(0, 4).map(x => `${x.node} ${x.days}/${days}d`).join(', ')}${short.length > 4 ? `, +${short.length - 4} more` : ''}`
+    : '';
+  return `showing ${what} from ${body.source}${gap}`;
+}
+
+/// Which part of the moment was just changed. The caller needs it to tell "the whole of a day" (an energy
+/// question) from "this instant of it" (a power one) — the two want different metrics, and guessing from
+/// the value alone cannot distinguish a freshly-picked day from a re-render.
+export type HistoryPick = 'day' | 'time' | 'span' | 'live';
+
+/// A "show this moment instead of now" control (#372). Returns the ISO instant to request, or '' for live.
+///
+/// Kept deliberately plain: a datetime input and a Live button. The value goes to the server as ?at=, and
+/// the page renders whatever it gets back — a historical view is the same diagram built from the values of
+/// that instant, not a second rendering path.
+export function historyControl(onChange: (what: HistoryPick) => void): {
+  row: HTMLElement, at: () => string, day: () => string, time: () => string, span: () => number,
+  setNote: (t: string) => void
+} {
+  const row = el('div', { class: 'ld-toolbar history-bar', style: { flexWrap: 'wrap', gap: '8px', margin: '0 0 8px' } });
+  // Separate date and time inputs, not a datetime-local: that control reports '' until BOTH halves are
+  // filled, so picking a day and leaving the time blank sent nothing and the page silently stayed live.
+  // Split, the time is genuinely optional — blank means the end of the chosen day.
+  const input = el('input', { type: 'date' }) as HTMLInputElement;
+  const timeIn = el('input', { type: 'time', step: '1' }) as HTMLInputElement;
+  const prev = btn('◀');
+  const next = btn('▶');
+  const live = btn('Live', 'primary');
+  // Which of the two things you are looking at, said plainly and in the same place every time. The date
+  // control alone does not answer it at a glance — an empty date reads as "not set yet" as easily as "now",
+  // and a diagram of last Tuesday looks exactly like a diagram of this second.
+  const badge = el('span', { class: 'pill good', text: 'LIVE' });
+  const note = el('span', { class: 'desc', style: { margin: '0' } });
+
+
+  live.title = 'Back to the current reading';
+
+  const today = () => new Date().toLocaleDateString('en-CA');   // yyyy-mm-dd in local time
+  const spanDays = () => Math.max(1, Number(spanSel.value) || 1);
+
+  // The arrows move by whatever is being shown: a day at a time on a single day, a week at a time on a
+  // week, a month on a month. Stepping one day through a 30-day window means 30 presses to see the window
+  // before it, and every one of those presses is a fresh set of history queries.
+  const step = (dir: number) => {
+    const from = input.value || today();
+    const d = new Date(from + 'T12:00:00');   // midday, so a DST shift cannot land on the previous day
+    d.setDate(d.getDate() + dir * spanDays());
+    const iso = d.toLocaleDateString('en-CA');
+    input.value = iso > today() ? today() : iso;   // no future days: there is nothing recorded there
+    onChange('day');
+  };
+
+  input.onchange = () => onChange('day');
+  timeIn.onchange = () => onChange('time');
+  prev.onclick = () => step(-1);
+  next.onclick = () => step(1);
+  const stepLabel = () => { const n = spanDays(); return n === 1 ? 'day' : n === 7 ? 'week' : `${n} days`; };
+  live.onclick = () => { input.value = ''; timeIn.value = ''; spanSel.value = '1'; syncSpan(); note.textContent = ''; onChange('live'); };
+
+  // The picker exists only if there is a backend to read from. With History switched off, a date control
+  // whose every answer is "history is turned off" is worse than no control at all.
+  //
+  // Read live rather than at build time, and re-read on every tab switch, so turning the feature on under
+  // Features brings the picker out without a reload.
+  const historyOn = () => !!((state.data && state.data.History) || {}).Enabled;
+  const syncEnabled = () => {
+    const on = historyOn();
+    row.classList[on ? 'remove' : 'add']('is-hidden');
+    // A day still selected when the feature is switched off has to stop being requested, or the page goes
+    // on asking for a moment that nothing can answer.
+    if (!on && input.value) { input.value = ''; timeIn.value = ''; spanSel.value = '1'; note.textContent = ''; onChange('live'); }
+  };
+
+  // One control rather than five: the arrows and inputs share a border and only the outer corners round,
+  // so the group reads as a single thing instead of a row of loose buttons.
+  const group = el('div', { class: 'input-group' }, prev, input, timeIn, next);
+
+  // How much of the past to add up. A week is the sum of seven daily totals — they all re-base at the same
+  // moment, so they add; a lifetime counter and an instantaneous power reading do not, which is why the
+  // server refuses a span for anything but the daily total and this offers one only alongside a date.
+  const spanSel = el('select', { title: 'Add up the daily totals over this many days, ending on the chosen day.' }) as HTMLSelectElement;
+  [['1', 'that day'], ['7', '7 days to it'], ['30', '30 days to it']]
+    .forEach(([v, t]) => spanSel.appendChild(el('option', { value: v, text: t })));
+  spanSel.onchange = () => { syncSpan(); onChange('span'); };
+
+  // A time within the day says nothing about a week of them, so the two cannot both be set.
+  const syncSpan = () => {
+    prev.title = 'Previous ' + stepLabel();
+    next.title = 'Next ' + stepLabel();
+    const many = spanDays() > 1;
+    timeIn.disabled = many;
+    if (many) timeIn.value = '';
+    timeIn.title = many
+      ? 'Not used over a span of days — each day is counted whole.'
+      : 'Optional. Leave blank for the end of the day — the day’s complete totals.';
+  };
+
+  row.append(badge, el('span', { class: 'desc', style: { margin: '0' }, text: 'At:' }), group,
+    el('span', { class: 'desc', style: { margin: '0' }, text: 'covering' }), spanSel, live, note);
+  syncSpan();
+  syncEnabled();
+  window.addEventListener?.('rpdu:activate', syncEnabled);
+  return {
+    row,
+    /// The instant to ask for.
+    ///
+    /// With a time, that time on the chosen day. Without one, the end of the day: a daily total is only
+    /// complete then, and midnight would return the total a moment after it re-based, which is zero. Either
+    /// way an instant still ahead of now is clamped to now, because nothing is recorded past it.
+    at: () => {
+      if (!historyOn() || !input.value) return '';
+      const when = new Date(`${input.value}T${timeIn.value || '23:59:59'}`);
+      const now = new Date();
+      return (when > now ? now : when).toISOString();
+    },
+    day: () => (historyOn() ? input.value : ''),
+    time: () => timeIn.value,
+    /// Days to add up, ending on the chosen day. 1 is the plain "that moment" view.
+    span: () => (historyOn() && input.value ? spanDays() : 1),
+    setNote: (t: string) => {
+      note.textContent = t;
+      // The badge follows what was actually rendered, not what the picker says: a date that produced no
+      // answer leaves the live view on screen, and calling that historical would be a lie about the figures.
+      const past = !!t;
+      badge.className = 'pill ' + (past ? 'warn' : 'good');
+      badge.textContent = past ? 'HISTORICAL' : 'LIVE';
+      badge.title = past ? t : 'These are the latest readings.';
+    },
+  };
+}
+
+// The tag selected on the diagram, kept across redraws: the Sankey repaints on every live push, and a
+// highlight that cleared itself every few seconds would be unusable.
